+++ src/ctf01d/include/ctf01d_config.h
@@ -53,8 +53,11 @@ public:
   virtual bool apply_config() = 0;
   virtual const std::vector<ctf01d::service_config> &services() = 0;
   virtual const std::vector<ctf01d::team_config> &teams() = 0;
+  virtual bool contains_team_id(const std::string &team_id) const = 0;
+  virtual std::string find_team_id_by_subnet(const std::string &ip) const = 0;
   virtual int scoreboard_port() const = 0;
   virtual std::string scoreboard_html_folder() const = 0;
+  virtual std::shared_ptr<ctf01d::var_bool> scoreboard_auto_detection_team_id_by_subnet_ip() const = 0;
   virtual bool scoreboard_random() const = 0;
   virtual std::shared_ptr<ctf01d::var_bool> scoreboard_metrics_enabled() const = 0;
   virtual std::shared_ptr<ctf01d::var_allowed_ip> scoreboard_metrics_allowed_for() const = 0;